Migrate from Neon Postgres to Turso: dataset was 10-20x over Neon's free-tier cap
Bulk-syncing 42.5k rows (a fraction of the 8,502-company registry) hit
Neon's 512MB project limit mid-run and crashed several sync processes.
Root cause: raw_json + raw_html (never rendered in the UI, pure
provenance/future-proofing) accounted for ~60% of row storage. Dropped
both columns entirely rather than trim around the edges — projected
full-registry size (1.3-2.9GB) still wouldn't fit Neon free tier but
comfortably fits Turso's 5GB.

Schema ported from pg-core to sqlite-core (bigserial->autoincrement,
jsonb->text/json-mode, timestamp->integer unixepoch, numeric->real).
Full-text search (tsvector/GIN) has no SQLite equivalent, replaced with
multi-word substring AND matching. The Postgres regex-based US-location
region filter is replaced with a SQL substr/instr-based token extraction
(a 50-state OR-chain of LIKE patterns hit SQLite's 100-deep expression
tree limit) — same word lists from lib/geo.ts, verified against real data
to match the original heuristic's behavior.

Also hardens scripts/sync.ts: a secondary DB write failure (e.g. logging
a failed sync attempt) no longer crashes the whole batch run — this is
exactly what took down 5 of 7 platforms simultaneously when Neon filled up.

+++ scripts/seed-sync-state.ts
@@ -60,11 +60,11 @@ async function main() {
   }
 
   const [{ count: tier1 }] = await db
-    .select({ count: sql<number>`count(*)::int` })
+    .select({ count: sql<number>`count(*)` })
     .from(syncState)
     .where(sql`tier = 1`);
   const [{ count: total }] = await db
-    .select({ count: sql<number>`count(*)::int` })
+    .select({ count: sql<number>`count(*)` })
     .from(syncState);
 
   console.log(`done — ${inserted} sites processed, ${total} total in sync_state (${tier1} tier-1)`);
